Asteroids, EIN and Parallaxing
Asteroids:
- Added in all asteroid sprites
- Reworked spawning algorithm to spawn more consistantly

EIN:
- Added in an effect indicator that tells the players what effects EIN is having on the play
- Reworked the effects chance. now EIN has a chance toeffect the ship that increases/decreases in relation to how happy he is and how long the game hs lasted
- Fixed a bug with updating the health bar

Ship:
-Added a damage animation that plays when the ship is hit by an asteroid

Misc:
- Added parallaxing backgrounds
- Added the radar base sprite
- Adjusted the play area

+++ Assets/Scripts/Jeff/Healthbar.cs
@@ -5,7 +5,7 @@ public class Healthbar : MonoBehaviour
 {
     private GameObject textBox;
     public Text healthBar;
-    public float currentHealth;
+    public int currentHealth;
 
     [SerializeField]
     private Slider slider;
@@ -63,7 +63,7 @@ public class Healthbar : MonoBehaviour
             death();
         }
 
-        slider.value = currentHealth/4;
+        slider.value = (float)currentHealth/4;
     }
     public void death()
     {
